Removed CHANGELOG.md and transferred its contents to changelog.txt
In an attempt to standardize our documentation process and keep the repository clean and organized, CHANGELOG.md was removed and its contents were transferred to a newly created file, changelog.txt.

At the same time, README.md was simplified significantly. Extraneous sponsorships, projects, and massive features block were removed. The content was stripped back to just a brief but descriptive concept of what WP Kit Elementor is.

A new readme.txt file was also introduced to provide the necessary meta information for the WordPress theme repository.

This major cleanup is intended to create a more sustained, focused release documentation that would benefit both developers and users. It will also ensure the maintenance of the theme is efficient moving forward.