Update Author URI and theme description in style.scss
Author URI and Description for WPKit Elementor theme were updated in the style.scss file. The URI was changed from http:// to https:// to enhance security. At the same time, the Descriptions was updated to enhance readability and consistency throughout the file. This change is part of the ongoing efforts to ensure maximum security and easy comprehension of the theme documentations for both developers and users.