The given commit message does not correspond with the diff provided. Here's an appropriate commit message:
Commit message:
Update package name in sidebar.php documentation

Updated the documentation in sidebar.php file in the WP Kit Elementor theme. Previously, the package name 'WP Kit Elementor' had a space which was inconsistent with our naming conventions. It was changed to 'WPKitElementor' for better consistency  and to avoid potential issues due to spacing.