Rename package name and change main content wrapper in footer.php
The package name has been renamed from 'WP Kit Elementor' to 'WPKitElementor' to ensure convention consistency within the code. Also, the main content wrapper has been changed from a 'main' tag with the id 'main' to a 'div' with the id 'content'. This improves HTML semantics as the 'main' tag should enclose the main unique content of a specific page, not the entire website. The skip-to-content link in 'header.php' is also updated to target '#content' for better accessibility support for screen readers.