Update skip link target and change content wrapper in header.php
The commit modifies the 'header.php' file of the WP Kit Elementor theme. The target of 'skip-link' is changed to '#content' from '#main' to improve accessibility for screen-reader users. Additionally, the main content wrapper is now a 'div' with id 'content', instead of a 'main' element with id 'main'. This enhances HTML semantics to accurately represent the unique content of a particular page, not the entire website.