Update package.json with bundling and translation tasks
Updated the package.json file in the wpkit-elementor theme to include tasks for creating a bundle of the theme and improving the translation process. Added a "bundle" script that uses 'dir-archiver' to create a .zip file of the theme, excluding certain unnecessary files. This will facilitate simpler distribution of the theme. Also adjusted the "translate" scripts to make .pot and .json files in the 'languages' directory rather than 'resources/lang'. This change is in line with WordPress standards for translation file locations. Added a new dependency, 'dir-archiver'.