Extend page rendering in wpkit-elementor theme
Page.php in the wpkit-elementor theme is updated to enhance the page rendering. Now the pages display a title, featured image, and the content. In the case of post pagination, a label is added to indicate the 'Pages'. Furthermore, comments section is installed at the end of the page if comments are open or already present. The motivation behind these changes is to increase page comprehensibility, provide more information for featured image and comments section, and make the pagination system more user-friendly.