Revert theme version to 1.0.0 and refactor header implementation
In this commit, the version of the theme in both style.css and style.scss files has been changed back to 1.0.0 from 1.0.2. The code related to the site's header, which was previously located in a separate file (header.php), is now directly included in the main header file for simpler file structure. Also, the readme text has been simplified and the package.json file has been updated for improved exclusion handling in bundle script.