Added editor style support to wp-kit-elementor theme
Added support for 'search-form', 'script', and 'style' to the html5 theme support in the functions.php file of the wp-kit-elementor theme. Also added the 'style-editor.css' file and the 'style-editor.scss' file. Integrated the processing of 'style-editor.scss' into 'webpack.mix.js'. The 'style-editor.css' and 'style-editor.scss' files contain instructions not to modify them directly and to use a custom directory for changes.