Upgrade theme requirement and refine copyright details
Updated both the WordPress version requirement for WP Kit Elementor and the copyright details in style.css and style.scss files. Notably, the theme's copyright was expanded with GPL distribution terms and the required WordPress version was increased to ensure better compatibility and enhanced user experience with the up-to-date WordPress version.