Correct product name reference in readme
This commit corrects the product name reference from "WP Kit Elementor" to "WpKit Elementor" in the theme's readme.txt file. It also includes the creation of a new README.md file with similar information for GitHub users. Lastly, the style.css update for WP Kit Elementor theme is indicated in the changelog.