Update color variable and references
A light-gray color variable has been added to _variables.scss and different sections of our code have been updated to use this new color variable instead of the hardcoded color value. The changes were made in _comments.scss and _posts-and-pages.scss to ensure uniformity in the color scheme across the Sass theme.