Update dotenv.config path in Elementor script
This commit adjusts the path used for 'dotenv.config' in the update-version-in-files.js script. The script is part of the wp-kit-elementor theme, and changing the path ensures the script can access the '.env' configuration file correctly. This change is critical for the proper functioning of the script for version updating tasks.